Crate moonbeam

Moonbase parachain collator

Modules

chain_spec
Moonbeam Chain Specifications and utilities for building them.

cli
Moonbeam CLI Library. Built with structopt

command
This module constructs and executes the appropriate service components for the given subcommand

inherents
This module is responsible for building the inherent data providers that a Moonbeam node needs.

rpc
A collection of node-specific RPC methods.

service
This module assembles the Moonbeam service components, executes them, and manages communication between them. This is the backbone of the client-side node implementation.

Functions

main