Module moonbeam::cli

Moonbeam CLI Library. Built with structopt

This module defines the Moonbeam node's Command Line Interface (CLI) It is built using structopt and inherits behavior from Substrate's sc_cli crate.

Structs

BuildSpecCommand

Cli

ExportGenesisStateCommand
Command for exporting the genesis state of the parachain

ExportGenesisWasmCommand
Command for exporting the genesis wasm file.

RelayChainCli

RunCmd

Enums

Sealing
Block authoring scheme to be used by the dev service.

Subcommand
Sub-commands supported by the collator.

Functions

parse_h160