Struct moonbeam::cli::RelayChainCli

pub struct RelayChainCli {
    pub base: RunCmd,
    pub chain_id: Option<String>,
    pub base_path: Option<PathBuf>,
}

Fields
base: RunCmd

The actual relay chain cli object.

chain_id: Option<String>

Optional chain id that should be passed to the relay chain.

base_path: Option<PathBuf>

The base path that should be used by the relay chain.

Implementations

impl RelayChainCli

pub fn new<'a>(
base_path: Option<PathBuf>,
chain_id: Option<String>,
relay_chain_args: impl Iterator<Item = &'a String>
) -> Self

Create a new instance of Self.