Module moonbeam::inherents

This module is responsible for building the inherent data providers that a Moonbeam node needs.

A service builder can call the build_inherent_data_providers function to get the providers the node needs based on the parameters it passed in.

This module also includes a MOCK inherent data provider for the validataion data inherent. This mock provider provides stub data that does not represent anything "real" about the external world, but can pass the runtime's checks. This is useful in testing for example, running the --dev service without a relay chain backbone.

Structs

MockValidationDataInherentDataProvider
Inherent data provider that supplies mocked validation data.

Functions

build_inherent_data_providers
Build the inherent data providers for the node.