Module moonbeam::service

This module assembles the Moonbeam service components, executes them, and manages communication between them. This is the backbone of the client-side node implementation.

This module can assemble: PartialComponents: For maintence tasks without a complete node (eg import/export blocks, purge) Full Service: A complete parachain node including the pool, rpc, network, embedded relay chain Dev Service: A leaner service without the relay chain backing.

Structs

Executor
A unit struct which implements NativeExecutionDispatch feeding in the hard-coded runtime.

NativeExecutor
A generic CodeExecutor implementation that uses a delegate to determine wasm code equivalence and dispatch to native code when possible, falling back on WasmExecutor when not.

Functions

new_dev
Builds a new development service. This service uses manual seal, and mocks the parachain inherent.

new_partial
Starts a ServiceBuilder for a full service.

start_node
Start a normal parachain node.

start_node_impl
Start a node with the given parachain Configuration and relay chain Configuration.

Type Definitions

FullBackend

FullClient