Function moonbeam::service::new_partial

pub fn new_partial(
    config: &Configuration, 
    author: Option<H160>, 
    mock_inherents: bool
) -> Result<PartialComponents<TFullClient<Block, RuntimeApi, Executor>, TFullBackend<Block>, (), BasicQueue<Block, PrefixedMemoryDB<BlakeTwo256>>, FullPool<Block, TFullClient<Block, RuntimeApi, Executor>>, (FrontierBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, Executor>>, TFullClient<Block, RuntimeApi, Executor>>, PendingTransactions, Option<FilterPool>)>, ServiceError>

Starts a ServiceBuilder for a full service.

Use this macro if you don't actually need the full service, but just the builder in order to be able to perform chain operations.