Function moonbeam::service::start_node_impl

async fn start_node_impl<RB>(
    parachain_config: Configuration, 
    collator_key: CollatorPair, 
    author_id: Option<H160>, 
    polkadot_config: Configuration, 
    id: Id, 
    collator: bool, 
    _rpc_ext_builder: RB
) -> Result<(TaskManager, Arc<TFullClient<Block, RuntimeApi, Executor>>)> where
    RB: Fn(Arc<TFullClient<Block, RuntimeApi, Executor>>) -> IoHandler<Metadata> + Send + 'static, 

Start a node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.